requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pillow
tensorflow
twilio
python-multipart
opencv-python
imagehash
ultralytics
numpy